Add files via upload
Client server sends a GPX file(representing the tracking route of the athlete) to master node. Master node is capable to serve and contact multiple workers (multithreaded) and send each worker chunks of N waypoints in a Round Robin order. Every worker executes the requests at the same time (multithreaded) and push back the intermediate results to master. Master reduces to the final result and forwards them to the mobile application. For the backend system we used a dummy server (client server). 
Master - worker use TCP sockets for communication